requirement.txt
bs4
lxml
json
re
urllib
python-slugify